Make AlignmentMode return algGermanPolar rather than algPolar because this was confusing the ASCOM TheSkyX Telescope plug-in, which in turn was not informing TheSkyX that the mount was a German mount. This caused TPoint not to be able to correctly model both sides of the sky simultaneously.
Make SideOfPier return PropertyNotImplementedExceptions for Get and Set instead of returning pierSide.Unknown when configured not to report pier side. This was causing the ASCOM TheSkyX Telescope plug-in to report that the Gemini driver could return  pier side even when it was configured not to do so. (The plug-in will also be modified to prevent other drivers falling into this trap)